refac: move MessageProducer to common package
