+++ core/lerna.json
@@ -1,0 +1,6 @@
+{
+  "packages": [
+    "services/**"
+  ],
+  "version": "0.0.0"
+}
