+++ core/services/eslint-config/package.json
@@ -1,5 +1,13 @@
 {
   "name": "@diva/eslint-config",
   "version": "0.1.0",
-  "main": "index.js"
+  "main": "index.js",
+  "devDependencies": {
+    "eslint": "^7.31.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-prettier": "^3.4.0",
+    "prettier": "^2.3.2"
+  }
 }
