+++ core/services/user-management/.dockerignore
@@ -34,7 +34,7 @@ bower_components
 build/Release
 
 # Dependency directories
-node_modules/
+node_modules
 jspm_packages/
 
 # Optional npm cache directory
