+++ core/services/user-management/package.json
@@ -16,7 +16,6 @@
     "@diva/common": "^1.0.0",
     "ajv": "^8.0.0",
     "ajv-formats": "^2.0.0",
-    "asyncapi-validator": "^2.5.0",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
     "chalk": "^4.1.0",
@@ -26,7 +25,6 @@
     "hasha": "^5.2.2",
     "jsondiffpatch": "^0.4.1",
     "jsonwebtoken": "^8.5.1",
-    "kafkajs": "^1.15.0",
     "lodash": "^4.17.21",
     "mongodb": "^3.6.4",
     "multer": "^1.4.2",
@@ -37,16 +35,11 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@diva/eslint-config": "^0.1.0",
     "chai": "^4.3.4",
     "chai-openapi-response-validator": "^0.13.0",
     "chai-sorted": "^0.2.0",
-    "eslint": "^7.22.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.1",
     "mocha": "^8.3.2",
-    "prettier": "^2.2.1",
     "rewire": "^5.0.0",
     "supertest": "^6.1.3"
   }
