+++ core/services/user-management/test/testApp.js
@@ -1,29 +1,30 @@
-const boot = require("../server");
+const messagesProducerService = require("@diva/common/MessageProducer");
+const boot = require("@diva/common/expressServer");
 const { passport } = require("../utils/passport");
 const { db } = require("../utils/database");
 const usersRouter = require("../routes/users");
 const userImagesRouter = require("../routes/userImages");
-const messagesProducerService = require("../services/MessagesProducerService");
-const { loadAsyncAPISpec } = require("../utils/validation/messagesValidation");
 const { loadSchemas } = require("../utils/validation/jsonSchemaValidation");
 
 const port = 0;
 
 const createTestServer = async () =>
-  boot(async (app) => {
-    app.use(passport.initialize());
+  boot(
+    async (app) => {
+      app.use(passport.initialize());
 
-    app.use("/users", usersRouter);
-    app.use("/userImages", userImagesRouter);
+      app.use("/users", usersRouter);
+      app.use("/userImages", userImagesRouter);
 
-    await Promise.all([
-      db.connect("userTestDb"),
-      messagesProducerService.init(() =>
-        console.log("Faked Sending message in test mode")
-      ),
-      loadSchemas(),
-      loadAsyncAPISpec(),
-    ]);
-  }, port);
+      await Promise.all([
+        db.connect("userTestDb"),
+        messagesProducerService.init(() =>
+          console.log("Faked Sending message in test mode")
+        ),
+        loadSchemas(),
+      ]);
+    },
+    { port }
+  );
 
 module.exports = createTestServer;
